(PC-32696)[PRO] refactor: mv IndividualOfferConfirmation to new arch folder
