+++ pro/src/pages/IndividualOffer/IndividualOfferConfirmation/components/IndividualOfferConfirmationScreen.module.scss
