+++ pro/src/pages/IndividualOffer/IndividualOfferConfirmation/components/IndividualOfferConfirmationScreen.tsx
